read usb power status on PB0
this ditches the status LED in favor of an additional input with which
we can read the state of the USB signal from the powerboost. this
allows us to detect when external power is available.

adds an additional POWERWAIT state because the USB signal can be
erroneously high when things first power up.