teach pipower about STATE_UNMANAGED
A long press in STATE_IDLE2 will now transition to STATE_UNMANAGED.
In this state, pipower will not respond to signals from the Pi nor
will it monitor the USB power. In this state, a short press of the
power button will toggle the current power state.  A long press
returns to STATE_IDLE2.